backend/requirements.txt
django>=5.1,<7
djangorestframework>=3.15,<4
Pillow>=10.0,<13
django-cors-headers>=4.3,<5
